requirements.txt
flask
requests
beautifulsoup4
anthropic
python-docx
gunicorn
